+++ sdk/python/pulumi_cloudamqp/integration_metric.py
@@ -24,87 +24,18 @@ class IntegrationMetric(pulumi.CustomResource):
                  name: Optional[pulumi.Input[str]] = None,
                  private_key: Optional[pulumi.Input[str]] = None,
                  project_id: Optional[pulumi.Input[str]] = None,
+                 queue_allowlist: Optional[pulumi.Input[str]] = None,
                  queue_whitelist: Optional[pulumi.Input[str]] = None,
                  region: Optional[pulumi.Input[str]] = None,
                  secret_access_key: Optional[pulumi.Input[str]] = None,
                  tags: Optional[pulumi.Input[str]] = None,
+                 vhost_allowlist: Optional[pulumi.Input[str]] = None,
                  vhost_whitelist: Optional[pulumi.Input[str]] = None,
                  __props__=None,
                  __name__=None,
                  __opts__=None):
         """
-<<<<<<< HEAD
-        This resource allows you to create and manage, forwarding metrics to third party integrations for a CloudAMQP instance. Once configured, the metrics produced will be forward to corresponding integration.
-
-        Only available for dedicated subscription plans.
-
-        ## Argument references
-
-        The following arguments are supported:
-
-        * `name`              - (Required) The name of the third party log integration. See `Integration service reference`
-        * `region`            - (Optional) Region hosting the integration service.
-        * `access_key_id`     - (Optional) AWS access key identifier.
-        * `secret_access_key` - (Optional) AWS secret access key.
-        * `api_key`           - (Optional) The API key for the integration service.
-        * `email`             - (Optional) The email address registred for the integration service.
-        * `project_id`        - (Optional) The project identifier.
-        * `private_key`       - (Optional) The private access key.
-        * `client_email`      - (Optional) The client email registered for the integration service.
-        * `tags`              - (Optional) Tags. e.g. env=prod, region=europe.
-        * `queue_whitelist`   - (Optional) Whitelist queues using regular expression. Leave empty to include all queues.
-        * `vhost_whitelist`   - (Optional) Whitelist vhost using regular expression. Leave empty to include all vhosts.
-
-        This is the full list of all arguments. Only a subset of arguments are used based on which type of integration used. See Integration type reference below for more information.
-
-        ## Integration service references
-
-        Valid names for third party log integration.
-
-        | Name          | Description |
-        |---------------|---------------------------------------------------------------|
-        | cloudwatch    | Create an IAM with programmatic access. |
-        | cloudwatch_v2 | Create an IAM with programmatic access. |
-        | datadog       | Create a Datadog API key at app.datadoghq.com |
-        | datadog_v2    | Create a Datadog API key at app.datadoghq.com
-        | librato       | Create a new API token (with record only permissions) here: https://metrics.librato.com/tokens |
-        | newrelic      | Deprecated! |
-        | newrelic_v2   | Find or register an Insert API key for your account: Go to insights.newrelic.com > Manage data > API keys. |
-        | stackdriver   | Create a service account and add 'monitor metrics writer' role, then download credentials. |
-
-        ## Integration type reference
-
-        Valid arguments for third party log integrations.
-
-        Required arguments for all integrations: *name*<br>
-        Optional arguments for all integrations: *tags*, *queue_whitelist*, *vhost_whitelist*
-
-        | Name | Type | Required arguments |
-        | ---- | ---- | ---- |
-        | Cloudwatch             | cloudwatch     | region, access_key_id, secret_access_key |
-        | Cloudwatch v2          | cloudwatch_v2  | region, access_key_id, secret_access_key |
-        | Datadog                | datadog        | api_key, region |
-        | Datadog v2             | datadog_v2     | api_key, region |
-        | Librato                | librato        | email, api_key |
-        | New relic (deprecated) | newrelic       | - |
-        | New relic v2           | newrelic_v2    | api_key, region |
-        | Stackdriver            | stackdriver    | project_id, private_key, client_email |
-
-        ## Dependency
-
-        This resource depends on CloudAMQP instance identifier, `cloudamqp_instance.instance.id`.
-
-        ## Import
-
-        `cloudamqp_integration_metric`can be imported using the name argument of the resource together with CloudAMQP instance identifier. The name and identifier are CSV separated, see example below.
-
-        ```sh
-         $ pulumi import cloudamqp:index/integrationMetric:IntegrationMetric <resource_name> <name>,<instance_id>`
-        ```
-
-=======
         Create a IntegrationMetric resource with the given unique name, props, and options.
->>>>>>> 5ed2468... Upgrae to v1.8.3 of the CloudAMQP Terraform Provider
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[str] access_key_id: AWS access key identifier. (Cloudwatch)
@@ -116,11 +47,13 @@ class IntegrationMetric(pulumi.CustomResource):
         :param pulumi.Input[str] name: The name of metrics integration
         :param pulumi.Input[str] private_key: The private key. (Stackdriver)
         :param pulumi.Input[str] project_id: Project ID. (Stackdriver)
-        :param pulumi.Input[str] queue_whitelist: (optional) whitelist using regular expression
+        :param pulumi.Input[str] queue_allowlist: (optional) allowlist using regular expression
+        :param pulumi.Input[str] queue_whitelist: **Deprecated**
         :param pulumi.Input[str] region: AWS region for Cloudwatch and [US/EU] for Data dog/New relic. (Cloudwatch, Data Dog, New Relic)
         :param pulumi.Input[str] secret_access_key: AWS secret key. (Cloudwatch)
         :param pulumi.Input[str] tags: (optional) tags. E.g. env=prod,region=europe
-        :param pulumi.Input[str] vhost_whitelist: (optional) whitelist using regular expression
+        :param pulumi.Input[str] vhost_allowlist: (optional) allowlist using regular expression
+        :param pulumi.Input[str] vhost_whitelist: **Deprecated**
         """
         if __name__ is not None:
             warnings.warn("explicit use of __name__ is deprecated", DeprecationWarning)
@@ -150,10 +83,18 @@ class IntegrationMetric(pulumi.CustomResource):
             __props__['name'] = name
             __props__['private_key'] = private_key
             __props__['project_id'] = project_id
+            __props__['queue_allowlist'] = queue_allowlist
+            if queue_whitelist is not None:
+                warnings.warn("""use queue_allowlist instead""", DeprecationWarning)
+                pulumi.log.warn("queue_whitelist is deprecated: use queue_allowlist instead")
             __props__['queue_whitelist'] = queue_whitelist
             __props__['region'] = region
             __props__['secret_access_key'] = secret_access_key
             __props__['tags'] = tags
+            __props__['vhost_allowlist'] = vhost_allowlist
+            if vhost_whitelist is not None:
+                warnings.warn("""use vhost_allowlist instead""", DeprecationWarning)
+                pulumi.log.warn("vhost_whitelist is deprecated: use vhost_allowlist instead")
             __props__['vhost_whitelist'] = vhost_whitelist
         super(IntegrationMetric, __self__).__init__(
             'cloudamqp:index/integrationMetric:IntegrationMetric',
@@ -174,10 +115,12 @@ class IntegrationMetric(pulumi.CustomResource):
             name: Optional[pulumi.Input[str]] = None,
             private_key: Optional[pulumi.Input[str]] = None,
             project_id: Optional[pulumi.Input[str]] = None,
+            queue_allowlist: Optional[pulumi.Input[str]] = None,
             queue_whitelist: Optional[pulumi.Input[str]] = None,
             region: Optional[pulumi.Input[str]] = None,
             secret_access_key: Optional[pulumi.Input[str]] = None,
             tags: Optional[pulumi.Input[str]] = None,
+            vhost_allowlist: Optional[pulumi.Input[str]] = None,
             vhost_whitelist: Optional[pulumi.Input[str]] = None) -> 'IntegrationMetric':
         """
         Get an existing IntegrationMetric resource's state with the given name, id, and optional extra
@@ -195,11 +138,13 @@ class IntegrationMetric(pulumi.CustomResource):
         :param pulumi.Input[str] name: The name of metrics integration
         :param pulumi.Input[str] private_key: The private key. (Stackdriver)
         :param pulumi.Input[str] project_id: Project ID. (Stackdriver)
-        :param pulumi.Input[str] queue_whitelist: (optional) whitelist using regular expression
+        :param pulumi.Input[str] queue_allowlist: (optional) allowlist using regular expression
+        :param pulumi.Input[str] queue_whitelist: **Deprecated**
         :param pulumi.Input[str] region: AWS region for Cloudwatch and [US/EU] for Data dog/New relic. (Cloudwatch, Data Dog, New Relic)
         :param pulumi.Input[str] secret_access_key: AWS secret key. (Cloudwatch)
         :param pulumi.Input[str] tags: (optional) tags. E.g. env=prod,region=europe
-        :param pulumi.Input[str] vhost_whitelist: (optional) whitelist using regular expression
+        :param pulumi.Input[str] vhost_allowlist: (optional) allowlist using regular expression
+        :param pulumi.Input[str] vhost_whitelist: **Deprecated**
         """
         opts = pulumi.ResourceOptions.merge(opts, pulumi.ResourceOptions(id=id))
 
@@ -214,10 +159,12 @@ class IntegrationMetric(pulumi.CustomResource):
         __props__["name"] = name
         __props__["private_key"] = private_key
         __props__["project_id"] = project_id
+        __props__["queue_allowlist"] = queue_allowlist
         __props__["queue_whitelist"] = queue_whitelist
         __props__["region"] = region
         __props__["secret_access_key"] = secret_access_key
         __props__["tags"] = tags
+        __props__["vhost_allowlist"] = vhost_allowlist
         __props__["vhost_whitelist"] = vhost_whitelist
         return IntegrationMetric(resource_name, opts=opts, __props__=__props__)
 
@@ -294,10 +241,18 @@ class IntegrationMetric(pulumi.CustomResource):
         return pulumi.get(self, "project_id")
 
     @property
+    @pulumi.getter(name="queueAllowlist")
+    def queue_allowlist(self) -> pulumi.Output[Optional[str]]:
+        """
+        (optional) allowlist using regular expression
+        """
+        return pulumi.get(self, "queue_allowlist")
+
+    @property
     @pulumi.getter(name="queueWhitelist")
     def queue_whitelist(self) -> pulumi.Output[Optional[str]]:
         """
-        (optional) whitelist using regular expression
+        **Deprecated**
         """
         return pulumi.get(self, "queue_whitelist")
 
@@ -326,10 +281,18 @@ class IntegrationMetric(pulumi.CustomResource):
         return pulumi.get(self, "tags")
 
     @property
+    @pulumi.getter(name="vhostAllowlist")
+    def vhost_allowlist(self) -> pulumi.Output[Optional[str]]:
+        """
+        (optional) allowlist using regular expression
+        """
+        return pulumi.get(self, "vhost_allowlist")
+
+    @property
     @pulumi.getter(name="vhostWhitelist")
     def vhost_whitelist(self) -> pulumi.Output[Optional[str]]:
         """
-        (optional) whitelist using regular expression
+        **Deprecated**
         """
         return pulumi.get(self, "vhost_whitelist")